backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
openai-whisper==20231117
torch==2.1.1
transformers==4.36.2
accelerate==0.25.0 